+++ src/frequenz/sdk/timeseries/formulas/_parser.py
@@ -18,6 +18,7 @@ from frequenz.sdk.timeseries._base_types import QuantityT
 
 from . import _ast, _token
 from ._base_ast_node import AstNode
+from ._exceptions import FormulaSyntaxError
 from ._formula import Formula
 from ._functions import FunCall, Function
 from ._lexer import Lexer
@@ -65,11 +66,17 @@ class _Parser(Generic[QuantityT]):
     ):
         """Initialize the parser."""
         self._name: str = name
+        self._formula: str = formula
         self._lexer: Peekable[_token.Token] = Peekable(Lexer(formula))
         self._telemetry_fetcher: ResampledStreamFetcher = telemetry_fetcher
         self._create_method: Callable[[float], QuantityT] = create_method
 
     def _parse_term(self) -> AstNode[QuantityT] | None:
+        """Parse a term.
+
+        A term is any expression. It may contain binary operators,
+        bracketed expressions, function calls, literals and component IDs.
+        """
         factor = self._parse_factor()
         if factor is None:
             return None
@@ -80,20 +87,27 @@ class _Parser(Generic[QuantityT]):
             next_factor = self._parse_factor()
 
             if next_factor is None:
-                raise ValueError(
-                    f"Expected factor after operator at span: {token.span}"
+                raise FormulaSyntaxError(
+                    formula=self._formula,
+                    span=(token.span[1], token.span[1] + 1),
+                    message="Expected expression",
                 )
 
             if isinstance(token, _token.Plus):
-                factor = _ast.Add(span=token.span, left=factor, right=next_factor)
+                factor = _ast.Add(left=factor, right=next_factor)
             elif isinstance(token, _token.Minus):
-                factor = _ast.Sub(span=token.span, left=factor, right=next_factor)
+                factor = _ast.Sub(left=factor, right=next_factor)
 
             token = self._lexer.peek()
 
         return factor
 
     def _parse_factor(self) -> AstNode[QuantityT] | None:
+        """Parse a factor.
+
+        A factor is any expression that does not contain
+        addition or subtraction outside of parentheses.
+        """
         unary = self._parse_unary()
 
         if unary is None:
@@ -104,29 +118,40 @@ class _Parser(Generic[QuantityT]):
             token = next(self._lexer)
             next_unary = self._parse_unary()
             if next_unary is None:
-                raise ValueError(f"Expected unary after operator at span: {token.span}")
+                raise FormulaSyntaxError(
+                    formula=self._formula,
+                    span=(token.span[1], token.span[1] + 1),
+                    message="Expected expression",
+                )
 
             if isinstance(token, _token.Mul):
-                unary = _ast.Mul(span=token.span, left=unary, right=next_unary)
+                unary = _ast.Mul(left=unary, right=next_unary)
             elif isinstance(token, _token.Div):
-                unary = _ast.Div(span=token.span, left=unary, right=next_unary)
+                unary = _ast.Div(left=unary, right=next_unary)
 
             token = self._lexer.peek()
 
         return unary
 
     def _parse_unary(self) -> AstNode[QuantityT] | None:
+        """Parse unary.
+
+        A unary is any expression that does not contain any binary
+        operators outside of parentheses.
+        """
         token: _token.Token | None = self._lexer.peek()
         if token is not None and isinstance(token, _token.Minus):
             token = next(self._lexer)
             primary: AstNode[QuantityT] | None = self._parse_primary()
             if primary is None:
-                raise ValueError(
-                    f"Expected primary expression after unary '-' at position {token.span}"
+                raise FormulaSyntaxError(
+                    formula=self._formula,
+                    span=(token.span[1], token.span[1] + 1),
+                    message="Expected expression",
                 )
 
-            zero_const = _ast.Constant(span=token.span, value=self._create_method(0.0))
-            return _ast.Sub(span=token.span, left=zero_const, right=primary)
+            zero_const = _ast.Constant(value=self._create_method(0.0))
+            return _ast.Sub(left=zero_const, right=primary)
 
         return self._parse_primary()
 
@@ -136,11 +161,19 @@ class _Parser(Generic[QuantityT]):
 
         expr: AstNode[QuantityT] | None = self._parse_term()
         if expr is None:
-            raise ValueError(f"Expected expression after '(' at position {oparen.span}")
+            raise FormulaSyntaxError(
+                formula=self._formula,
+                span=(oparen.span[1], oparen.span[1] + 1),
+                message="Expected expression",
+            )
 
         token: _token.Token | None = self._lexer.peek()
         if token is None or not isinstance(token, _token.CloseParen):
-            raise ValueError(f"Expected ')' after expression at position {expr.span}")
+            raise FormulaSyntaxError(
+                formula=self._formula,
+                span=oparen.span,
+                message="Unmatched parenthesis",
+            )
 
         _ = next(self._lexer)  # consume ')'
 
@@ -148,40 +181,70 @@ class _Parser(Generic[QuantityT]):
 
     def _parse_function_call(self) -> AstNode[QuantityT] | None:
         fn_name: _token.Token = next(self._lexer)
+        function_class: type[Function[QuantityT]] | None = (
+            Function.function_class_by_name(fn_name.value)
+        )
+
+        if function_class is None:
+            raise FormulaSyntaxError(
+                formula=self._formula,
+                span=fn_name.span,
+                message="Unknown function name",
+            )
+
         params: list[AstNode[QuantityT]] = []
 
         token: _token.Token | None = self._lexer.peek()
         if token is None or not isinstance(token, _token.OpenParen):
-            raise ValueError(
-                f"Expected '(' after function name at position {fn_name.span}"
+            raise FormulaSyntaxError(
+                formula=self._formula,
+                span=(fn_name.span[1], fn_name.span[1] + 1),
+                message="Expected '(' after function name",
             )
+        oparen = next(self._lexer)  # consume '('
 
-        _ = next(self._lexer)  # consume '('
         while True:
             param = self._parse_term()
             if param is None:
-                raise ValueError(
-                    f"Expected argument in function call at position {fn_name.span}"
+                raise FormulaSyntaxError(
+                    formula=self._formula,
+                    span=(token.span[1], token.span[1] + 1),
+                    message="Expected argument",
                 )
             params.append(param)
 
             token = self._lexer.peek()
-            if token is not None and isinstance(token, _token.Comma):
+            if token is None:
+                raise FormulaSyntaxError(
+                    formula=self._formula,
+                    span=oparen.span,
+                    message="Unmatched parenthesis",
+                )
+
+            if isinstance(token, _token.Comma):
                 _ = next(self._lexer)  # consume ','
                 continue
-            if token is not None and isinstance(token, _token.CloseParen):
+            if isinstance(token, _token.CloseParen):
                 _ = next(self._lexer)  # consume ')'
                 break
-            raise ValueError(
-                f"Expected ',' or ')' in function call at position {fn_name.span}"
+
+            raise FormulaSyntaxError(
+                formula=self._formula,
+                span=token.span,
+                message="Expected ',' or ')'",
             )
 
-        return FunCall(
-            span=fn_name.span,
-            function=Function.from_string(fn_name.value, params),
-        )
+        return FunCall(function=function_class(params))
 
     def _parse_primary(self) -> AstNode[QuantityT] | None:
+        """Parse a primary.
+
+        A primary(expression) can be any of the following:
+        - A number literal
+        - A component ID
+        - A function call
+        - A bracketed expression
+        """
         token: _token.Token | None = self._lexer.peek()
         if token is None:
             return None
@@ -194,7 +257,6 @@ class _Parser(Generic[QuantityT]):
         if isinstance(token, _token.Component):
             _ = next(self._lexer)  # consume token
             comp = _ast.TelemetryStream(
-                span=token.span,
                 source=f"#{token.id}",
                 metric_fetcher=make_component_stream_fetcher(
                     self._telemetry_fetcher, ComponentId(int(token.id))
@@ -205,9 +267,7 @@ class _Parser(Generic[QuantityT]):
 
         if isinstance(token, _token.Number):
             _ = next(self._lexer)
-            return _ast.Constant(
-                span=token.span, value=self._create_method(float(token.value))
-            )
+            return _ast.Constant(value=self._create_method(float(token.value)))
 
         if isinstance(token, _token.OpenParen):
             return self._parse_bracketed()
@@ -220,7 +280,20 @@ class _Parser(Generic[QuantityT]):
     def parse(self) -> Formula[QuantityT]:
         expr = self._parse_term()
         if expr is None:
-            raise ValueError("Empty formula.")
+            raise FormulaSyntaxError(
+                formula=self._formula,
+                span=None,
+                message="Empty formula",
+            )
+        # There should not be any tokens left
+        token = self._lexer.peek()
+        if token is not None:
+            raise FormulaSyntaxError(
+                formula=self._formula,
+                span=token.span,
+                message="Unexpected token",
+            )
+
         return Formula(
             name=self._name,
             root=expr,